+++ src/modules/rlm_cache/rlm_cache.h
@@ -217,7 +217,7 @@ typedef cache_status_t	(*cache_entry_set_ttl_t)(rlm_cache_config_t const *config
  *	#cache_acquire_t callback was provided.
  * @return number of entries in the cache.
  */
-typedef uint32_t	(*cache_entry_count_t)(rlm_cache_config_t const *config, void *instance,
+typedef uint64_t	(*cache_entry_count_t)(rlm_cache_config_t const *config, void *instance,
 					       request_t *request, void *handle);
 
 /** Acquire a handle to access the cache
